Organize components folder into atoms, molecules and pages
