+++ frontend/src/App.tsx
@@ -3,15 +3,15 @@ import { Provider } from 'react-redux';
 import { ChakraProvider } from '@chakra-ui/react';
 import { Route, BrowserRouter as Router, Switch } from 'react-router-dom';
 import store from './state/store';
-import ErrorModal from './components/ErrorModal';
-import Navbar from './components/Navbar';
-import Footer from './components/Footer';
-import Frontpage from './pages/Frontpage';
+import ErrorModal from './components/atoms/ErrorModal';
+import Navbar from './components/atoms/Navbar';
+import Footer from './components/atoms/Footer';
+import Frontpage from './components/pages/Frontpage';
 import './css/App.css';
-import About from './pages/About';
-import OntologyPage from './pages/OntologyPage';
+import About from './components/pages/About';
+import OntologyPage from './components/pages/OntologyPage';
 
-const App = () => (
+const App: React.FC = () => (
   <ChakraProvider>
     <Provider store={store}>
       <div className="App">
