+++ frontend/src/components/atoms/Footer.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+import { Flex } from '@chakra-ui/react';
+
+const Footer = () => (
+  <Flex bgGradient="linear(to-r, green.200, pink.500)" w="100%" p={4} color="white" />
+);
+
+export default Footer;
