+++ frontend/src/components/pages/About.tsx
