+++ frontend/src/components/pages/Frontpage.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import SustainabilityGoals from '../molecules/SustainabilityGoalsList';
+
+const Frontpage: React.FC = () => <SustainabilityGoals />;
+
+export default Frontpage;
